requirements.txt
langchain
streamlit
python-dotenv
pypdf
openai
tiktoken
faiss-cpu
duckduckgo-search
llama-index
span-marker